pyproject.toml
[project]
name = "papermerge"
version = "3.6.0"
description = "Open source document management system for digital archives"
authors = [
    { name = "Eugen Ciur", email = "eugen@papermerge.com" }
]
maintainers = [
    { name = "Eugen Ciur", email = "eugen@papermerge.com" }
]
license = "Apache-2.0"
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3"
]
keywords = [
    "PDF",
    "metadata",
    "custom fields",
    "PDF/A",
    "scanning",
    "dms",
    "document-management",
    "paperless",
    "archives",
    "records"
]
requires-python = ">=3.13"
dependencies = [
    "redis",
    "celery[gevent,redis]",
    "pyyaml",
    "python-magic",
    "uritemplate",
    "pdf2image",
    "uvicorn",
    "python-jose",
    "img2pdf",
    "passlib",
    "sqlalchemy",
    "typer",
    "rich",
    "pytz",
    "pydantic-settings",
    "alembic",
    "fastapi[standard]",
    "taskipy",
    "asyncpg",
    "aiofiles",
    "psycopg2-binary",
    "prefect>=3.0,<4.0",
    "simpleeval>=1.0",
    "httpx>=0.27",
    "uuid-extension",
    "blake3",
    "prometheus-client",
]

[project.urls]
Homepage = "https://www.papermerge.com/"
Documentation = "https://docs.papermerge.io/"
Repository = "https://github.com/papermerge/papermerge-core"
Issues = "https://github.com/ciur/papermerge/issues"

[dependency-groups]
dev = [
    "freezegun>=1.5.5",
    "pytest",
    "pytest-asyncio",
]

[project.optional-dependencies]
cloud_storage = [
    "botocore",
    "boto3",
    "cryptography",
]
quality = [
    "numpy>=2.0",
    "pillow>=11.0",
]
search = [
    "elasticsearch>=8.0",
    "meilisearch>=0.30",
]
scanner = [
    "zeroconf>=0.132",
    "httpx>=0.27",
]
nlp = [
    "spacy>=3.8",
]
all = [
    "papermerge[cloud_storage,quality,search,scanner,nlp]",
]


[project.scripts]
pm = "papermerge.cli:main"
docs = "papermerge.core.cli.docs:app"
sign_url = "papermerge.storage.cli.sign_url:app"


[tool.taskipy.settings]
use_vars = true

[tool.taskipy.variables]
env_vars = "PM_API_PREFIX=''"

[tool.taskipy.tasks]
test = { cmd = "{env_vars} pytest", cwd = "." }
worker = "celery -A config worker -c 5"
server = "fastapi dev papermerge/app.py"
migrate = "alembic upgrade head"

[tool.ruff]
target-version = "py311"
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "F", # pyflakes
    "I001", # isort
    "UP", # pyupgrade
]
ignore = [
    "UP007"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["papermerge"]

[tool.pytest.ini_options]
testpaths = ["papermerge"]
asyncio_mode = "auto"
